add(compiler/ssa)[feat]: `PreDefinedMethods` for reserving a symbol in the LLVM IR
In code generation (especially when compiling large Rust programs with many crates), function
definitions might refer to each other in arbitrary order. To handle this correctly, tidec performs
codegen in two passes for each function:
1. Pre-definition (this method): Reserve space in the LLVM module for the function, declare it with proper attributes, and make it globally available to other definitions.
2. Definition (done later): Actually emit the LLVM IR body of the function.
